+++ programs/templatetags/form_tags.py
@@ -192,3 +192,18 @@ def get_relationship(adult, student):
         ).relationship_to_student
     except AdultStudentRelationship.DoesNotExist:
         return ""
+
+@register.filter
+def multiply(value, arg):
+    try:
+        return float(value) * float(arg)
+    except (ValueError, TypeError):
+        return 0
+
+
+@register.filter
+def divide(value, arg):
+    try:
+        return float(value) / float(arg)
+    except (ValueError, TypeError, ZeroDivisionError):
+        return 0